Removed the cyclical dependency  (#1519)
