+++ src/config.js
@@ -287,3 +287,11 @@ export const PLANS_MAPPING = {
 
 
 export const DEFAULT_SETTING_KEEP_ALL_WORKSPACES_LOADED = false;
+
+export const DEFAULT_SERVICE_LIMIT = 3;
+
+export const DEFAULT_TODOS_WIDTH = 300;
+export const TODOS_MIN_WIDTH = 200;
+export const DEFAULT_TODOS_VISIBLE = true;
+export const DEFAULT_IS_FEATURE_ENABLED_BY_USER = true;
+export const TODOS_PARTITION_ID = 'persist:todos';
