+++ src/features/basicAuth/index.js
@@ -1,21 +1,12 @@
 import { ipcRenderer } from 'electron';
-import { observable } from 'mobx';
 
 import BasicAuthComponent from './Component';
 
+import { state as ModalState } from './store';
+
 const debug = require('debug')('Ferdi:feature:basicAuth');
 
-const defaultState = {
-  isModalVisible: true,
-  service: null,
-  authInfo: null,
-};
-
-export const state = observable(defaultState);
-
-export function resetState() {
-  Object.assign(state, defaultState);
-}
+const state = ModalState;
 
 export default function initialize() {
   debug('Initialize basicAuth feature');
@@ -38,21 +29,6 @@ export function mainIpcHandler(mainWindow, authInfo) {
   mainWindow.webContents.send('feature:basic-auth-request', {
     authInfo,
   });
-}
-
-export function sendCredentials(user, password) {
-  debug('Sending credentials to main', user, password);
-
-  ipcRenderer.send('feature-basic-auth-credentials', {
-    user,
-    password,
-  });
-}
-
-export function cancelLogin() {
-  debug('Cancel basic auth event');
-
-  ipcRenderer.send('feature-basic-auth-cancel');
 }
 
 export const Component = BasicAuthComponent;
