+++ src/features/delayApp/constants.js
@@ -1,0 +1,6 @@
+import { DEFAULT_FEATURES_CONFIG } from '../../config';
+
+export const config = {
+  delayOffset: DEFAULT_FEATURES_CONFIG.needToWaitToProceedConfig.delayOffset,
+  delayDuration: DEFAULT_FEATURES_CONFIG.needToWaitToProceedConfig.wait,
+};
