+++ src/features/delayApp/index.js
@@ -1,16 +1,13 @@
 import { autorun, observable, reaction } from 'mobx';
 import moment from 'moment';
 import DelayAppComponent from './Component';
-
+import { config as configurations } from './constants';
 import { DEFAULT_FEATURES_CONFIG } from '../../config';
 import { getUserWorkspacesRequest } from '../workspaces/api';
 
 const debug = require('debug')('Ferdi:feature:delayApp');
 
-export const config = {
-  delayOffset: DEFAULT_FEATURES_CONFIG.needToWaitToProceedConfig.delayOffset,
-  delayDuration: DEFAULT_FEATURES_CONFIG.needToWaitToProceedConfig.wait,
-};
+export const config = configurations;
 
 export const state = observable({
   isDelayAppScreenVisible: DEFAULT_FEATURES_CONFIG.needToWaitToProceed,
