+++ src/features/quickSwitch/index.js
@@ -1,14 +1,9 @@
-import { observable } from 'mobx';
+import { state as ModalState } from './store';
 
 export { default as Component } from './Component';
+const state = ModalState;
 
 const debug = require('debug')('Ferdi:feature:quickSwitch');
-
-const defaultState = {
-  isModalVisible: false,
-};
-
-export const state = observable(defaultState);
 
 export default function initialize() {
   debug('Initialize quickSwitch feature');
