+++ src/features/quickSwitch/store.js
@@ -1,0 +1,7 @@
+import { observable } from 'mobx';
+
+const defaultState = {
+  isModalVisible: false,
+};
+
+export const state = observable(defaultState);
