+++ src/features/serviceLimit/index.js
@@ -3,8 +3,6 @@ import { ServiceLimitStore } from './store';
 
 const debug = require('debug')('Ferdi:feature:serviceLimit');
 
-export const DEFAULT_SERVICE_LIMIT = 3;
-
 let store = null;
 
 export const serviceLimitStore = new ServiceLimitStore();
