+++ src/features/shareFranz/store.js
@@ -1,0 +1,8 @@
+import { observable } from 'mobx';
+
+const defaultState = {
+  isModalVisible: false,
+  lastShown: null,
+};
+
+export const state = observable(defaultState);
