+++ src/features/todos/constants.js
@@ -2,3 +2,7 @@ export const IPC = {
   TODOS_HOST_CHANNEL: 'TODOS_HOST_CHANNEL',
   TODOS_CLIENT_CHANNEL: 'TODOS_CLIENT_CHANNEL',
 };
+
+export const TODOS_ROUTES = {
+  TARGET: '/todos',
+};
