+++ src/features/todos/index.js
@@ -5,16 +5,6 @@ const debug = require('debug')('Ferdi:feature:todos');
 
 export const GA_CATEGORY_TODOS = 'Todos';
 
-export const DEFAULT_TODOS_WIDTH = 300;
-export const TODOS_MIN_WIDTH = 200;
-export const DEFAULT_TODOS_VISIBLE = true;
-export const DEFAULT_IS_FEATURE_ENABLED_BY_USER = true;
-export const TODOS_PARTITION_ID = 'persist:todos';
-
-export const TODOS_ROUTES = {
-  TARGET: '/todos',
-};
-
 export const todosStore = new TodoStore();
 
 export default function initTodos(stores, actions) {
